Reject outbound header blocks with invalid headers
There are several rules we use to reject inbound header blocks as
invalid, but we don't enforce them on headers we're sending:

* A Transfer-Encoding header must have value "trailers".
* No blocks should have connection-specific headers.

These are rules mandated by RFC 7540. We already have the code to
police these rules, so apply it to outbound as well as inbound
headers.

Inbound headers always have byte strings for names/values, whereas
we may still have Unicode strings for the outbound headers (they're
encoded after this validation step), so we need to be tolerant
to Unicode headers here.